Do not use random live patch IDs
This patch replaces the use of a randomly generated number as live patch
ID. Now, the description file is used as input to the MD4 digest
algorithm, then the digest is used as ID. This makes live patch IDs
deterministic, which avoids reproducibility concerns.

The MD4 implementation is a copy of IETF's RFC 1320 [1] reference
implementation. It keeps its copyright and license notice, as required.

[1] https://tools.ietf.org/html/rfc1320