Improve the testing framework
The test suite currently has lots of duplicated code, mostly from three
major sources:

  1. boilerplate to start live patchable processes;
  2. logic to compare their output against expected patterns;
  3. boilerplate to use and diagnose ulp tools.

In order to reduce the duplication, this patch refactors the testing
framework so that it takes more advantage of python's programming model.
Class inheritance helps reduce the boilerplate, whereas exceptions help
reduce verification logic.

1. Process initialization

The test suite now defines a new module, testsuite, which redefines the
spawn class from pexpect, inheriting from it. The main purpose of the
redefinition is to provide default parameters that automatically convert
the spawned child into a live patchable process, as well as to
automatically request that all output be sent to stdout.

The new defaults reduce process initialization boilerplate from:

  child = pexpect.spawn('./' + testname, timeout=1, env=preload, encoding='utf-8')
  child.logfile = sys.stdout

to:

  child = testsuite.spawn('testname')

2. Verification logic

The main interface to pexpect, the expect method, takes a list of
expected strings, and returns an index to which of them has been found
in the child's output. However, when the expected patterns are not
found, pexpect prints lots of information (of which only some are
relevant to the test suite), then causes the test script to exit. The
new class overrides this method, extending its interface to include an
list of patterns to check but reject, as well as capturing unmatched
patterns and raising relevant exceptions, instead of printing internal
information.

With the overridden method, code blocks such as:

  child.sendline('dozen')
  index = child.expect(['13', '12']);
  print('Third call to libdozens... ', end='')
  if index == 0:
    print('ok.')
  if index == 1:
    print('not ok; old behavior.')
    exit (1)

become simply:

  child.expect('13', reject='12');

Since the new testing framework raises exceptions when a pattern is not
found, it is still possible to detect and diagnose these situation.
However, instead of building a list of expected patterns that include
EOF and TIMEOUT, then parsing the result, a try/expect pattern should be
used, such as:

  try:
    child.expect('pattern')
  except EOFError:
    print('Diagnostics message for EOF')
  except TimeoutError:
    print('Diagnostics message for timeout')

3. Live patching tools

Finally, running ulp tools required repetitive boilerplate to start them
and to diagnose their results, which were mostly left unchecked. For
instance, before commit ID 92567f45d59e, the trigger tool could fail
with SIGSEGV, but the test suite would hide it, because the return code
of subprocess.run calls were left untested, as the logs showed:

  ...
  ulp: entering the critical section (process hijacking).
  Failed to apply livepatch #1 for libparameters

which now properly raise exceptions to help debug test case failures:

  ulp: entering the critical section (process hijacking).
  Traceback (most recent call last):
    File "parameters.py", line 32, in <module>
      child.livepatch('libparameters_livepatch1.ulp')
    File "testsuite.py", line 204, in livepatch
      tool.check_returncode()
    File "/usr/lib64/python3.8/subprocess.py", line 448, in check_returncode
      raise CalledProcessError(self.returncode, self.args, self.stdout,
  subprocess.CalledProcessError:
    Command '['ulp_trigger', '-p', '2183', 'libparameters_livepatch1.ulp', '-v']'
    died with <Signals.SIGSEGV: 11>.

Moreover, testsuite.livepatch makes it easier to run the tools, reducing
the boilerplate from:

  ret = subprocess.run([trigger, '-r', '100', '-p', str(child.pid),
                        'libblocked_livepatch1.ulp'], timeout=1)
  if ret.returncode:
    print('Failed to apply livepatch #1 for libblocked')
    errors = 1

to:

  child.livepatch('libblocked_livepatch1.ulp', retries=100)