Validate input in test case
In the 'terminal' test case, two live patchable process are created: the
first of them, named 'parent' is started by the test script in the usual
fashion, i.e. with testsuite.spawn(); the second, named 'child', is
started by the parent via a fork syscall. After forking, the parent
prints the pid of its child to stdout. Afterwards, the test script reads
the printed pid to use as the --pid argument to the checker tool, which
expects a number in string format. Since the checker tool is external to
the python script (called with subprocess.run()), validate the sanity of
the argument.

From the 'redzone' test case, just remove the XXX annotation because
blocking while waiting for the child process to finish is not only OK
but actually required: OK because the timeout in testsuite.spawn() will
take care of not letting the test suite run forever, if the child never
stops; and required because the child process is designed to take a long
time to finish, so isalive() is useless.