Add support for unexported variables access in live patches
So far, none of the live patch examples in the test suite needs access
to data objects in the target library, instead, all of them replace
functions that only deal with local variables or function arguments
(either on the stack or in registers). However, direct access to data
objects in the target library is desirable and even required by some
live patches.

If a live patch needs to access a data object that has been exported by
the target library, i.e. a GLOBAL object, the compiler toolchain is able
to generate the relocations that the loader needs. However, when the
data object has not been exported, i.e. a LOCAL object, the toolchain is
unable to generate a reference and fails with an 'undefined reference'
error message.

To bridge this gap, this patch extends the live patch metadata with
information about LOCAL data objects. The information is comprised of
the name of the variable in the target library, the name of a reference
variable in the live patch, as well as the offsets that these two data
objects have within their DSOs. The following example (taken from the
new test case added by this patch) of a live patch description file
provides a clarification:

  1. /run/user/1001/libpulp/tests/.libs/libaccess_livepatch1.so
  2. @/run/user/1001/libpulp/tests/.libs/libaccess.so.0
  3. banner_set:new_banner_set
  4. #banner:ulpr_banner:0000000000004020:0000000000004038

Lines 1 and 2 are not new and inform Libpulp about the in-disk paths to
the live patch DSO and to the target library, respectively. Line 3 is
also not new, and provides a pair of replaced and replacement functions.
Line 4 is new. It is preceded by the '#' character and divided into 4
parts, in a predefined order: 1. the name of the data object in the
target library, 'banner'; 2. the name of another data object,
'ulpr_banner', which sits within the live patch object and is used as a
reference to 'banner'; 3. the offset of 'banner' from the loading
address of the target library; 4. the offset of 'ulpr_banner' from the
loading address of the live patch object.

Then, when a live patch gets applied, Libpulp uses the aforementioned
offsets and makes 'ulpr_banner' point to the loaded address of 'banner',
which makes it possible for the live patch to access the data object in
the target library.

PS: The names of the objects (in the target library and in the live
patch) are not really necessary and, in fact, not used by Libpulp.